+++ eve/tests/methods/patch.py
@@ -13,6 +13,41 @@ from eve.tests.utils import DummyEvent
 
 
 class TestPatch(TestBase):
+    def test_patch_not_override_other_fields(self):
+        self.app.config["ENFORCE_IF_MATCH"] = False
+        # create a data
+        r, status = self.post(self.test_patch_url, data={"name": "name", "contact": {}})
+        self.assert201(status)
+        # check the data is created correctly
+        data, status = self.get(self.test_patch, item=r["_id"])
+        self.assert200(status)
+
+        # patch the data
+        _, status = self.patch(
+            self.test_patch_url + "/" + data["_id"], data={"contact.phone": "new_phone"}
+        )
+        self.assert200(status)
+        # other fields should not be touched
+        data, status = self.get(self.test_patch, item=r["_id"])
+        self.assert200(status)
+        self.assertEqual(data["name"], "name")
+        self.assertTrue("contact" in data)
+        self.assertEqual(data["contact"]["phone"], "new_phone")
+        self.assertEqual(data["contact"]["email"], "default_email")
+
+        # patch other field of the data
+        _, status = self.patch(
+            self.test_patch_url + "/" + data["_id"], data={"contact.email": "new_email"}
+        )
+        self.assert200(status)
+        # other fields should not be touched
+        data, status = self.get(self.test_patch, item=r["_id"])
+        self.assert200(status)
+        self.assertEqual(data["name"], "name")
+        self.assertTrue("contact" in data)
+        self.assertEqual(data["contact"]["phone"], "new_phone")
+        self.assertEqual(data["contact"]["email"], "new_email")
+
     def test_patch_to_resource_endpoint(self):
         _, status = self.patch(self.known_resource_url, data={})
         self.assert405(status)
@@ -333,7 +368,7 @@ class TestPatch(TestBase):
         self.assertTrue("OK" in r[STATUS])
 
     def test_patch_x_www_form_urlencoded_number_serialization(self):
-        del (self.domain["contacts"]["schema"]["ref"]["required"])
+        del self.domain["contacts"]["schema"]["ref"]["required"]
         field = "anumber"
         test_value = 3.5
         changes = {field: test_value}
@@ -462,7 +497,7 @@ class TestPatch(TestBase):
 
     def test_patch_readonly_field_with_previous_document(self):
         schema = self.domain["contacts"]["schema"]
-        del (schema["ref"]["required"])
+        del schema["ref"]["required"]
 
         # disable read-only on the field so we can store a value which is
         # also different form its default value.
@@ -617,7 +652,7 @@ class TestPatch(TestBase):
         """
         # this will fail as dependent field is missing even in the
         # document we are trying to update.
-        del (self.domain["contacts"]["schema"]["dependency_field1"]["default"])
+        del self.domain["contacts"]["schema"]["dependency_field1"]["default"]
         changes = {"dependency_field2": "value"}
         r, status = self.patch(
             self.item_id_url, data=changes, headers=[("If-Match", self.item_etag)]